backend/requirements.txt
fastapi==0.115.2 
uvicorn[standard]==0.32.0 
langchain==0.3.4 
langchain-google-genai==2.0.1 
langgraph==0.2.37 
pydantic==2.9.2 
python-dotenv==1.0.1 
pyyaml==6.0.2